Surface API now included in every GS2!
Dear GS2 fans,

Because we love a seamless, silky smooth experience, we have included Surface API in the latest GameStation 2s.
This means that when developers utilise this well, you're in for a great menu, lagging so little it might even make up for server lag.

So, how does this magical lag-free experience work?
Well, let's quote the forums post.
"The Surface API is a really powerful drawing API for ComputerCraft. It uses the least amount of drawing calls possible to reduce lag and tearing."
How can devs use it? And how does the API work?
"A surface is a texture that you can draw to.
It has 3 layers: the character, background color and text color layer.
The data of those 3 layers are stored in one sequential buffer so it uses the least amount of RAM.
Surfaces are wrapped in tables with the width, height, buffer and functions in it.
There are many useful functions, such as: clearing the surface, drawing text, pixels, lines, rectangles, scaling surfaces and more!
You can also load and save surfaces with the NFP format for the built-in paint,
or the more advanced NFT format which also supports text for programs such as NPaintPro and Sketch.
This is a complete replacement for the build-in paintutils API and maybe even the term API."

(Trust me, it's as good as it sounds once you utilise it fully)

There is also a full list of the (many) functions at the forum post:
http://www.computercraft.info/forums2/index.php?/topic/22397-surface-api-162/