Redirect API Added
Official notes on redirect API:

"The redirect api has only one directly-called function, createRedirectBuffer, which creates and returns a new redirect buffer object, but that object has all the methods necessary to be used as a redirect target on basic and advanced computers, as well as a few extra methods for managing the buffer and drawing it's contents to the screen.

There are two ways to get the contents of a buffer to the screen. The first is explicitly calling buffer.blit, which allows you to copy all or part of the buffer to some position on the screen. The second us with the buffer.makeActive method, which allows you to map the buffer to some portion of the screen, which will cause anything written to the buffer to immediately draw to that area of the screen as well. The first is useful for implementing a form of back buffering, or for allowing switching between two or more screens of data without having to rebuild those screens each time you switch. The second is useful for splitting the screen into smaller areas, each of which can be written to independently. Each of these buffered areas would have it's own independent cursor position and active colors, as well as scrolling and clearing without affecting the rest of the screen.

Redirect buffers can certainly be used independently, but they're at their most when combined with goroutines. goroutines has built-in support for redirect targets (any redirect targets, not just those created by the redirect api), and by combining the two you can easily do things like create a chat application that has one coroutine reading input, mapped to a single line buffer at the bottom of the screen, while another coroutine using a buffer fitting the rest of the screen prints the scrolling chat history.

A note about the buffer objects: in order to be compatible with term.redirect, the buffer object methods are wrapped to supply the buffer object as a parameter automatically, so despite being lua objects, you call their methods with "." instead of ":"."